+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+* [App] Graticule layer can no longer be selected
+  [#218](https://gitlab.data.bas.ac.uk/MAGIC/SIIS/-/issues/218)
+
 ### Changed
 
 * Footprint selection supports multiple features